Run the usb's code as a limited user (but not on a read-only fs)
Under Debian (and thus Raspbian) Stretch, systemd-nspawn is broken
for read-only mounting of root filesystems from directories (see
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=876103).

While there is a fix in the version of systemd in backports, that
version then fails to operate correctly in other ways (it initially
fails to boot properly due to issues with raspi-copies-and-fills,
even after fixing that it doesn't correctly unmount things when a
device is removed).

Instead we achieve similar isolation by running as an unprivileged
user inside the container.

Fixes #8.