Switch to apt-get
apt-get is designed for use in scripts and has a stable output;
apt is designed for humans and tries to do things like progress
bars. While we don't actually need to parse the output, apt-get
is still the right thing to use here.

This also moves to using the longer form of arguments for clarity.

Fixes #20.