Add CI linting
This deliberately _doesn't_ use flake8-mypy as I'm finding that
mypy is quite slow, even on a small codebase such as this. In turn
that makes the editor experience suboptimal as the quick linter
feedback from just running flake8 on its own is preferable.

For clarity: my editor also separately runs mypy as well, so I don't
mind that we have two check commands to use. I'm open to adding a
wrapper which does both or editing my editor config to disable
flake8-mypy if others would prefer the checks be combined.