Decode the output returned by subprocess.check_output before printing it
`subprocess.check_output` returns a `bytes`. If this is logged directly
it is converted using `repr` and so appears all on one line, as in the
log extract in this comment:
https://github.com/sourcebots/robot-api/issues/49#issuecomment-362643552)

Decoding the output into a `str` should cause it to be displayed
properly. Unicode decoding errors are handled by inserting a
Unicode Replacement Character (U+FFFD) rather than raising an
exception.