Use process groups for user code process
After the addition of idfc, calling self.process.terminate() only
terminates the idfc process, not the script process that is started
by idfc nor the python process that is started by script. This is
bad, because yanking the USB while the user code is running is no
longer killing the user code.

This is resolved by ensuring the idfc process is put into a new
process group, and sending the signal to all processes in this
process group rather than just the root idfc process.